dashboards_api/requirements.txt
fastapi
uvicorn[standart]
sqlalchemy
pandas
numpy
python-jose[cryptography]
passlib[bcrypt]